+++ src/Foundation/Models/MasterProduct.php
@@ -14,6 +14,8 @@ declare(strict_types=1);
 
 namespace Vanilo\Foundation\Models;
 
+use Illuminate\Database\Eloquent\Casts\Attribute;
+use Illuminate\Support\Carbon;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
@@ -25,6 +27,10 @@ use Vanilo\Properties\Traits\HasPropertyValues;
 use Vanilo\Support\Traits\HasImagesFromMediaLibrary;
 use Vanilo\Taxes\Traits\BelongsToTaxCategory;
 
+/**
+ * @property-read int $units_sold
+ * @property-read null|Carbon $last_sale_at
+ */
 class MasterProduct extends BaseMasterProduct implements HasMedia
 {
     use BelongsToTaxCategory;
@@ -38,5 +44,19 @@ class MasterProduct extends BaseMasterProduct implements HasMedia
     public function registerMediaConversions(Media $media = null): void
     {
         $this->loadConversionsFromVaniloConfig();
+    }
+
+    protected function unitsSold(): Attribute
+    {
+        return Attribute::make(
+            get: fn () => $this->variants()->sum('units_sold'),
+        );
+    }
+
+    protected function lastSaleAt(): Attribute
+    {
+        return Attribute::make(
+            get: fn () => $this->variants()->max('last_sale_at'),
+        );
     }
 }